requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
tensorflow-cpu>=2.16.1
tf-keras
numpy<2.0.0
pillow==10.2.0
google-generativeai==0.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
